isolate grpc server from processor
