+++ its_hub/integration/ext_proc/processor.py
@@ -5,33 +5,24 @@ This gRPC service implements the Envoy External Processor protocol to intercept
 HTTP requests to LLM endpoints and apply inference-time scaling algorithms.
 """
 
-import argparse
-import asyncio
+# ruff: noqa: I001
 import json
 import logging
 import time
 import uuid
 
-try:
-    import grpc
-    from envoy.config.core.v3 import base_pb2
-    from envoy.service.ext_proc.v3 import external_processor_pb2 as ext_proc_pb2
-    from envoy.service.ext_proc.v3 import external_processor_pb2_grpc as ext_proc_grpc
-    from envoy.type.v3 import http_status_pb2
-
-    # Import proto after setting up path
-    import its_hub.integration.ext_proc.proto  # noqa: F401
-except ImportError:
-    grpc = None  # type: ignore[assignment]
+import grpc
+import its_hub.integration.ext_proc.proto  # noqa: F401
+from envoy.config.core.v3 import base_pb2
+from envoy.service.ext_proc.v3 import external_processor_pb2 as ext_proc_pb2
+from envoy.service.ext_proc.v3 import external_processor_pb2_grpc as ext_proc_grpc
+from envoy.type.v3 import http_status_pb2
 
 from its_hub.api import ITSRequestConfig
 from its_hub.core.gateway import ITSGateway
 
 _ITS_HEADER_PREFIX = "x-its-"
 
-logging.basicConfig(
-    level=logging.INFO, format="%(asctime)s - %(name)s - %(levelname)s - %(message)s"
-)
 logger = logging.getLogger(__name__)
 
 
@@ -50,38 +41,6 @@ def _preview_message_content(message: dict, limit: int = 160) -> str:
     if not preview:
         return "<empty>"
     return preview[:limit] + ("…" if len(preview) > limit else "")
-
-
-def _configure_logging(level_name: str) -> None:
-    """Adjust log levels for root and key ITS modules."""
-    numeric_level = getattr(logging, level_name.upper(), None)
-    if not isinstance(numeric_level, int):
-        raise ValueError(f"Invalid log level: {level_name}")
-
-    logging.getLogger().setLevel(numeric_level)
-    for logger_name in (
-        "its_hub.integration.ext_proc.processor",
-        "its_hub.core.gateway",
-        "its_hub.core.algorithms.self_consistency",
-    ):
-        logging.getLogger(logger_name).setLevel(numeric_level)
-
-
-def _parse_args() -> argparse.Namespace:
-    """Parse CLI arguments for the ext_proc server."""
-    parser = argparse.ArgumentParser(description="ITS Envoy External Processor")
-    parser.add_argument(
-        "--port",
-        type=int,
-        default=50051,
-        help="Port to bind the gRPC server (default: 50051)",
-    )
-    parser.add_argument(
-        "--log-level",
-        default="INFO",
-        help="Logging level (e.g., DEBUG, INFO, WARNING)",
-    )
-    return parser.parse_args()
 
 
 class ExternalProcessorService(ext_proc_grpc.ExternalProcessorServicer):
@@ -439,43 +398,3 @@ class ExternalProcessorService(ext_proc_grpc.ExternalProcessorServicer):
         await self.gateway.ashutdown()
 
 
-async def serve(port: int = 50051):
-    """Start the gRPC server."""
-    server = grpc.aio.server()
-
-    processor = ExternalProcessorService()
-
-    ext_proc_grpc.add_ExternalProcessorServicer_to_server(processor, server)
-    server.add_insecure_port(f"[::]:{port}")
-
-    logger.info("Starting External Processor on port %d", port)
-    await server.start()
-
-    logger.info("External Processor is ready to receive requests from Envoy")
-
-    try:
-        await server.wait_for_termination()
-    except KeyboardInterrupt:
-        logger.info("Received shutdown signal")
-        await processor.shutdown()
-        await server.stop(grace=5)
-
-
-def main():
-    """Entry point for the external processor service."""
-    if grpc is None:
-        print(
-            "Error: ext_proc dependencies not installed.\n"
-            "Install with: pip install 'its_hub[ext_proc]'"
-        )
-        raise SystemExit(1)
-    args = _parse_args()
-    _configure_logging(args.log_level)
-    try:
-        asyncio.run(serve(port=args.port))
-    except KeyboardInterrupt:
-        logger.info("Service stopped")
-
-
-if __name__ == "__main__":
-    main()
