+++ pyproject.toml
@@ -28,7 +28,7 @@ dependencies = [
 
 [project.scripts]
 # its-iaas removed for MVP - will be added back in [iaas] extra if needed
-envoy-grpc = "its_hub.integration.ext_proc.processor:main"
+envoy-grpc = "its_hub.integration.ext_proc.server:main"
 
 [tool.setuptools_scm]
 version_file = "its_hub/_version.py"
